% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_to_text.R
\name{binary_to_text}
\alias{binary_to_text}
\title{Recode variables 0-1 to No-Yes}
\usage{
binary_to_text(x)
}
\arguments{
\item{x}{A numeric vector containing binary values (0 or 1).}
}
\value{
A character vector where 0 is replaced with 'No', 1 is replaced with 'Yes',
and all other values are replaced with NA.
}
\description{
This function takes a numeric vector with binary values (0 and 1) and
converts it into a character vector with 'No' for 0 and 'Yes' for 1.
Any value other than 0 or 1 is converted to NA.
}
