% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_from_rowsums.R
\name{create_from_rowsums}
\alias{create_from_rowsums}
\title{Create a Binary Indicator from Row Sums}
\usage{
create_from_rowsums(df, col_names, risk_behavior)
}
\arguments{
\item{df}{A dataframe containing the columns to be transformed and summed.}

\item{col_names}{Vector of column names to be converted to numeric and summed.}

\item{risk_behavior}{The name to be assigned to the binary indicator column.}
}
\value{
A modified dataframe with the original columns replaced by a binary
indicator column, reflecting the presence (1) or absence (0) of the aggregated
risk behavior, based on the sum of the specified columns.
}
\description{
This function transforms selected columns in a dataframe to numeric,
calculates their row-wise sums, and then creates a binary indicator column.
The binary column is determined by whether the row sum is equal to or
greater than 1. It dynamically generates the names for the row sum and binary
columns based on input parameters.
}
\details{
The function performs the following steps:
\enumerate{
\item Converts the specified columns (col_names) to numeric.
\item Calculates the row-wise sum of these columns, handling NA values.
\item Creates a binary indicator column (risk_behavior) where 1 indicates
a row sum >= 1, and 0 indicates a row sum < 1.
\item Removes the original columns used for row sums, retaining only the binary
indicator and other existing dataframe columns.
}
}
