% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary_2021.R
\docType{data}
\name{dictionary_2021}
\alias{dictionary_2021}
\title{YRBS Dictionary 2021}
\format{
An object of class \code{look_for} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 237 rows and 7 columns.
}
\usage{
data(dictionary_2021)
}
\description{
This dataset serves as a variable dictionary for the raw 2021 YRBS dataset.
It was generated using the \code{bulkreadr} package and provides metadata on the variables included in the dataset,
including their names, labels, and types.
}
\details{
Youth Risk Behavior Surveillance System (YRBS) Dictionary - 2021

The dictionary helps users understand the structure of the dataset
by listing variable names and descriptions, making it easier to work with the raw data.
}
\keyword{datasets}
