% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Saving YRBS Data to Your Computer.R
\name{list_package_datasets}
\alias{list_package_datasets}
\title{List Available Datasets in a Package}
\usage{
list_package_datasets(package_name)
}
\arguments{
\item{package_name}{Character. The name of the package to check for datasets.}
}
\value{
A character vector containing the names of available datasets in the package.
If the package is not installed or contains no datasets, an informative message is displayed.
}
\description{
This function retrieves and lists all available datasets within a specified package.
}
\examples{
\dontrun{
list_package_datasets("datasets")
list_package_datasets("dissertationData")
}

}
