% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_yrbs.R
\name{load_yrbs}
\alias{load_yrbs}
\title{Load YRBS Data from the Package}
\usage{
load_yrbs(dataset_name)
}
\arguments{
\item{dataset_name}{Character. The name of the dataset to load (e.g., "raw2017").}
}
\value{
The requested dataset as a fully evaluated object.
}
\description{
This function retrieves a dataset from the package and returns it as a fully evaluated object,
preventing lazy loading (promise objects).
}
\examples{
\dontrun{
my_data <- load_yrbs("raw2017")
}

}
