% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw2015.R
\docType{data}
\name{raw2015}
\alias{raw2015}
\title{raw2015}
\format{
A tibble with 15,624 and 239 variables
}
\usage{
data(raw2015)
}
\description{
This is the imported raw version of the 2015 YRBS Data.
This data set monitors priority health risk behaviors that contribute
markedly to the leading causes of death, disability, and social problems
among youth and adults in the United States.

These broad categories include:
\itemize{
\item Behaviors that contribute to unintentional injuries and violence
\item Sexual behaviors that contribute to unintended pregnancy and
sexually transmitted diseases, including HIV infection
\item Alcohol and other drug use
\item Tobacco use
\item Unhealthy dietary behaviors
\item Inadequate physical activity
}
}
\details{
Youth Risk Behavior Surveillance System 2015
}
\keyword{datasets}
