% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw2017.R
\docType{data}
\name{raw2017}
\alias{raw2017}
\title{raw2017}
\format{
A tibble with 14,765 and 247 variables
}
\usage{
data(raw2017)
}
\description{
This is the imported raw version of the 2017 YRBS Data.
This data set monitors priority health risk behaviors that contribute
markedly to the leading causes of death, disability, and social problems
among youth and adults in the United States.
These behaviors, often include:

\describe{
Please refer to the documentation provided in the 2017 National
YRBS Data Users Guide.
\itemize{
\item{Behaviors that contribute to unintentional injuries and violence}
\item{Sexual behaviors that contribute to unintended pregnancy and
sexually transmitted diseases, including HIV infection}
\item{Alcohol and other drug use}
\item{Tobacco use}
\item{Unhealthy dietary behaviors}
\item{Inadequate physical activity}
}
}
}
\details{
Youth Risk Behavior Surveillance System 2017
}
\keyword{datasets}
