% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw2021.R
\docType{data}
\name{raw2021}
\alias{raw2021}
\title{raw2021}
\format{
A tibble with 17,232 and 237 variables
}
\usage{
data(raw2021)
}
\description{
This is the imported raw version of the 2021 YRBS Data.
This data set monitors priority health risk behaviors that contribute
markedly to the leading causes of death, disability, and social problems
among youth and adults in the United States.
These behaviors, often include:

Please refer to the documentation provided in the 2021 National
YRBS Data Users Guide.

These broad categories include:
\itemize{
\item Behaviors that contribute to unintentional injuries and violence
\item Sexual behaviors that contribute to unintended pregnancy and
sexually transmitted diseases, including HIV infection
\item Alcohol and other drug use
\item Tobacco use
\item Unhealthy dietary behaviors
\item Inadequate physical activity
}
}
\details{
Youth Risk Behavior Surveillance System 2021
}
\keyword{datasets}
