% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissertationData-package.R
\docType{package}
\name{dissertationData-package}
\alias{dissertationData}
\alias{dissertationData-package}
\title{dissertationData: Cleaned and combined YRBS 2015, 2017, 2019, 2021, and 2023 data}
\description{
This package contains cleaned datasets from the Youth Risk Behavior Surveillance System (YRBS) for the years 2015 and 2023. The central goal of this project is to facilitate research on adolescent suicide attempts, initially for my (Catalina) dissertation, but now for anyone committed to **evidence-based decision-making**!
}
\author{
\strong{Maintainer}: Catalina Cañizares \email{cc9082@nyu.edu} (\href{https://orcid.org/0000-0002-6854-5205}{ORCID})

Authors:
\itemize{
  \item Francisco Cardozo \email{foc9@miami.edu} (\href{https://orcid.org/0000-0002-1925-4954}{ORCID})
}

}
\keyword{internal}
