% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_risk_behaviors.R
\name{plot_risk_behaviors}
\alias{plot_risk_behaviors}
\title{Plot Risk Behaviors}
\usage{
plot_risk_behaviors(
  data,
  behavior_vars,
  suicide_var,
  plot_title,
  fill_colors = c("#24bccb", "#4e2d86")
)
}
\arguments{
\item{data}{A data frame containing the variables for plotting.}

\item{behavior_vars}{A vector of column names for risk behavior variables.}

\item{suicide_var}{The column name for the suicide attempt variable.}

\item{plot_title}{The title of the plot.}

\item{fill_colors}{A character vector of colors for the plot fill.}
}
\value{
A ggplot object.
}
\description{
Plot Risk Behaviors
}
